requirements.txt
django-model-utils>=2.0
# Additional requirements go here
faker>=0.7.0
